Modificaciones para el D5 Ampliación de Subvenciones (Campos prioritarios) Incidencias varias que se han detectado y corregido en los módulos ya desarrollados. Subimos aunque no estén finalizados su desarrollos los siguientes módulos: Convenios, Presupuestos y Tráfico

+++ API_CA/API_CORE/src/main/java/org/ciudadesabiertas/utils/DatasetSearch.java
@@ -50,7 +50,7 @@ public interface DatasetSearch<T>  {
 			numbers.add("int");
 			numbers.add("double");
 			numbers.add("long");
-			numbers.add("long");	
+			numbers.add("float");	
 		}
